feat: migrate financial core to Long cents and four-market formatting
Store bill amounts and monthly budgets as minor units, restrict
SupportedCurrency to USD/CAD/MXN/EUR, and route every displayed
amount through CurrencyFormatter while parsing form input with
CurrencyParser. Room is bumped to version 10 with destructive
fallback for this pre-production schema.

+++ app/src/main/java/com/artie/chargemenot/data/local/UserSettingsEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class UserSettingsEntity(
     @PrimaryKey
     val id: Int = SETTINGS_ID,
-    val monthlyBudget: Double,
+    val monthlyBudget: Long,
     val isNagModeEnabled: Boolean = false,
     val selectedCurrency: String = com.artie.chargemenot.domain.model.UserSettings.DEFAULT_CURRENCY,
     val isOnboardingComplete: Boolean = false,
