+++ app/src/main/java/com/artie/chargemenot/domain/model/ForecastResult.kt
@@ -1,7 +1,7 @@
 package com.artie.chargemenot.domain.model
 
 data class ForecastResult(
-    val predictedAmount: Double,
+    val predictedAmount: Long,
     val percentageVariance: Double,
     val weatherStatus: WeatherStatus,
     val timelinePoints: List<ForecastTimelinePoint>,
@@ -10,7 +10,7 @@ data class ForecastResult(
 
 data class ForecastTimelinePoint(
     val monthLabel: String,
-    val amount: Double,
+    val amount: Long,
     val isPrediction: Boolean,
     val weatherStatus: WeatherStatus
 )
